# contracts/host/openapi.json
{
  "components": {
    "schemas": {
...
      "AddContextReferenceResult": {
        "additionalProperties": false,
        "properties": {
          "dataRevision": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          },
          "reference": {
            "$ref": "#/components/schemas/ContextReferenceSummary"
          }
        },
        "required": [
          "reference",
          "dataRevision"
        ],
        "type": "object"
      },
...
      "CreateWorkingContextResult": {
        "additionalProperties": false,
        "properties": {
          "context": {
            "$ref": "#/components/schemas/WorkingContextSummary"
          },
          "dataRevision": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "context",
          "dataRevision"
        ],
        "type": "object"
      },
...
      "ListWorkingContextsResult": {
        "additionalProperties": false,
        "properties": {
          "contexts": {
            "items": {
              "$ref": "#/components/schemas/WorkingContextSummary"
            },
            "type": "array"
          },
          "dataRevision": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          },
          "nextCursor": {
            "maxLength": 512,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "contexts",
          "dataRevision"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "correlationId": {
            "maxLength": 160,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "parameters": {
            "additionalProperties": false,
            "properties": {
              "currentRevision": {
                "maximum": 9007199254740991,
                "minimum": 0,
                "type": "integer"
              },
              "expectedRevision": {
                "maximum": 9007199254740991,
                "minimum": 0,
                "type": "integer"
              }
            },
            "type": "object"
          },
          "retryable": {
            "type": "boolean"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "correlationId",
          "retryable",
          "parameters"
        ],
        "type": "object"
      },
...
      "SetWorkScopeResumeResult": {
        "anyOf": [
          {
            "additionalProperties": false,
            "properties": {
              "area": {
                "enum": [
                  "manage"
                ],
                "type": "string"
              },
              "dataRevision": {
                "maximum": 9007199254740991,
                "minimum": 0,
                "type": "integer"
              },
              "resume": {
                "$ref": "#/components/schemas/ManagementResume"
              }
            },
            "required": [
              "area",
              "resume",
              "dataRevision"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "area": {
                "enum": [
                  "analyze"
                ],
                "type": "string"
              },
              "dataRevision": {
                "maximum": 9007199254740991,
                "minimum": 0,
                "type": "integer"
              },
              "resume": {
                "$ref": "#/components/schemas/AnalysisResume"
              }
            },
            "required": [
              "area",
              "resume",
              "dataRevision"
            ],
            "type": "object"
          }
        ]
      },
...
      "WorkingContextSummary": {
        "additionalProperties": false,
        "properties": {
          "analysisResumeVersion": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "type": "integer"
          },
          "boundary": {
            "maxLength": 2000,
            "minLength": 1,
            "type": "string"
          },
          "contextId": {
            "pattern": "^[1-9]\\d{0,15}$",
            "type": "string"
          },
          "contextVersion": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "type": "integer"
          },
          "createdAt": {
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d{1,3})?Z$",
            "type": "string"
          },
          "displayName": {
            "maxLength": 160,
            "minLength": 1,
            "type": "string"
          },
          "lifecycle": {
            "anyOf": [
              {
                "enum": [
                  "active"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "archived"
                ],
                "type": "string"
              }
            ]
          },
          "managementResumeVersion": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "type": "integer"
          },
          "nextStep": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string"
          },
          "pinnedOrder": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          },
          "purpose": {
            "maxLength": 2000,
            "minLength": 1,
            "type": "string"
          },
          "referenceCount": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          },
          "updatedAt": {
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d{1,3})?Z$",
            "type": "string"
          }
        },
        "required": [
          "contextId",
          "displayName",
          "lifecycle",
          "contextVersion",
          "referenceCount",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "WorkingContextWorkspace": {
        "additionalProperties": false,
        "properties": {
          "analysisResume": {
            "$ref": "#/components/schemas/AnalysisResume"
          },
          "context": {
            "$ref": "#/components/schemas/WorkingContextSummary"
          },
          "dataRevision": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          },
          "managementResume": {
            "$ref": "#/components/schemas/ManagementResume"
          },
          "references": {
            "items": {
              "$ref": "#/components/schemas/ContextReferenceSummary"
            },
            "type": "array"
          }
        },
        "required": [
          "context",
          "references",
          "dataRevision"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "hostBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Plysmith Application Host",
    "version": "0.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/working-contexts": {
      "get": {
        "operationId": "ListWorkingContexts",
        "parameters": [
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "pattern": "^(?:[1-9]|[1-9]\\d|100)$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "maxLength": 512,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListWorkingContextsResult"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          }
        }
      },
      "post": {
        "operationId": "CreateWorkingContext",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "boundary": {
                    "maxLength": 2000,
                    "minLength": 1,
                    "type": "string"
                  },
                  "displayName": {
                    "maxLength": 160,
                    "minLength": 1,
                    "type": "string"
                  },
                  "nextStep": {
                    "maxLength": 500,
                    "minLength": 1,
                    "type": "string"
                  },
                  "purpose": {
                    "maxLength": 2000,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "displayName"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateWorkingContextResult"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          }
        }
      }
    },
    "/working-contexts/{contextId}": {
      "get": {
        "operationId": "GetWorkingContextWorkspace",
        "parameters": [
          {
            "in": "path",
            "name": "contextId",
            "required": true,
            "schema": {
              "pattern": "^[1-9]\\d{0,15}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkingContextWorkspace"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          }
        }
      }
    },
    "/working-contexts/{contextId}/references": {
      "post": {
        "operationId": "AddContextReference",
        "parameters": [
          {
            "in": "path",
            "name": "contextId",
            "required": true,
            "schema": {
              "pattern": "^[1-9]\\d{0,15}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "anchorId": {
                    "pattern": "^[1-9]\\d{0,15}$",
                    "type": "string"
                  },
                  "itemId": {
                    "pattern": "^[1-9]\\d{0,15}$",
                    "type": "string"
                  }
                },
                "required": [
                  "itemId",
                  "anchorId"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddContextReferenceResult"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          }
        }
      }
    },
    "/working-contexts/{contextId}/resume": {
      "put": {
        "operationId": "SetWorkScopeResume",
        "parameters": [
          {
            "in": "path",
            "name": "contextId",
            "required": true,
            "schema": {
              "pattern": "^[1-9]\\d{0,15}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "additionalProperties": false,
                    "properties": {
                      "area": {
                        "enum": [
                          "manage"
                        ],
                        "type": "string"
                      },
                      "expectedResumeVersion": {
                        "anyOf": [
                          {
                            "maximum": 9007199254740991,
                            "minimum": 1,
                            "type": "integer"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "presentation": {
                        "anyOf": [
                          {
                            "enum": [
                              "list"
                            ],
                            "type": "string"
                          },
                          {
                            "enum": [
                              "atlas"
                            ],
                            "type": "string"
                          }
                        ]
                      },
                      "selectedAnchorId": {
                        "pattern": "^[1-9]\\d{0,15}$",
                        "type": "string"
                      },
                      "selectedItemId": {
                        "pattern": "^[1-9]\\d{0,15}$",
                        "type": "string"
                      }
                    },
                    "required": [
                      "area",
                      "expectedResumeVersion",
                      "presentation"
                    ],
                    "type": "object"
                  },
                  {
                    "additionalProperties": false,
                    "properties": {
                      "anchorId": {
                        "pattern": "^[1-9]\\d{0,15}$",
                        "type": "string"
                      },
                      "area": {
                        "enum": [
                          "analyze"
                        ],
                        "type": "string"
                      },
                      "expectedResumeVersion": {
                        "anyOf": [
                          {
                            "maximum": 9007199254740991,
                            "minimum": 1,
                            "type": "integer"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "itemId": {
                        "pattern": "^[1-9]\\d{0,15}$",
                        "type": "string"
                      },
                      "mode": {
                        "anyOf": [
                          {
                            "enum": [
                              "analyze"
                            ],
                            "type": "string"
                          },
                          {
                            "enum": [
                              "edit_inventory"
                            ],
                            "type": "string"
                          },
                          {
                            "enum": [
                              "edit_overlay"
                            ],
                            "type": "string"
                          }
                        ]
                      },
                      "revisionId": {
                        "pattern": "^[1-9]\\d{0,15}$",
                        "type": "string"
                      }
                    },
                    "required": [
                      "area",
                      "expectedResumeVersion",
                      "mode"
                    ],
                    "type": "object"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetWorkScopeResumeResult"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          }
        }
      }
    }
  },
  "security": [
    {
      "hostBearer": []
    }
  ]
}